Resolved some issues
