+++ API/appsettings.json
@@ -7,7 +7,7 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "DefaultConnection": "Server=localhost,3306;Database=medicalink_new;User=root;Password=;"
+    "DefaultConnection": "Server=localhost,3306;Database=medicalink_new;User=root;Password=root;"
   },
   "Jwt": {
     "Key": "MedicaLinkBackendSuperKey12457jklinjgjhiohio87946565ghfrsesgchggj",
